+++ src/app/app.module.ts
@@ -1,22 +1,11 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { ChartsModule } from 'ng2-charts';
-import { Routes, RouterModule } from '@angular/router';
-import { PatientModule } from './pages/patient/patient.module';
-import { DoctorModule } from './pages/doctor/doctor.module';
-import { AdminModule } from './pages/admin/admin.module';
 import { SharedModule } from './pages/shared/shared.module';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
-const routes: Routes = [
-  {
-    path: '',
-    redirectTo: 'patient/home',
-    pathMatch: 'full'
-  },
-];
 
 @NgModule({
   declarations: [
@@ -26,11 +15,7 @@ const routes: Routes = [
     BrowserModule,
     AppRoutingModule,
     ChartsModule,
-    PatientModule,
-    DoctorModule,
-    AdminModule,
     SharedModule,
-    RouterModule.forRoot(routes)
   ],
   providers: [],
   bootstrap: [AppComponent],
