+++ src/app/components/confirm-dialog/confirm-dialog.component.css
@@ -2,11 +2,12 @@
     font-weight: bold;
 }
 .confirm-dialog{
-    position: absolute;
+    position: fixed;
     width: 100%;
     height: 100%;
     top: 0;
     left: 0;
+    z-index: 99;
 }
 .background{
     position: absolute;
