+++ src/app/components/update-schedule/update-schedule.component.css
@@ -1,10 +1,11 @@
 
 .confirm-dialog{
-    position: absolute;
+    position: fixed;
     width: 100%;
     height: 100%;
     top: 0;
     left: 0;
+    z-index: 99;
 }
 .background{
     position: absolute;
