+++ src/app/components/update-schedule/update-schedule.component.html
@@ -27,7 +27,7 @@
             </div>
         </div>
         <div class="row">
-            <div class="col-lg-12">
+            <div class="col-lg-12" style="width: 100%; overflow: auto;">
                 <table>
                     <tr>
                         <th>
