+++ src/app/pages/shared/shared.module.ts
@@ -14,6 +14,11 @@ import { ProfileComponent } from './../../components/profile/profile.component';
 import { UnloginComponent } from './../../components/unlogin/unlogin.component';
 import { LoginDialogComponent } from './../../components/login-dialog/login-dialog.component';
 import { RegisterDialogComponent } from './../../components/register-dialog/register-dialog.component';
+import { ScheduleComponent } from './../../components/schedule/schedule.component';
+import { NotifDialogComponent } from './../../components/notif-dialog/notif-dialog.component';
+import { ConfirmDialogComponent } from './../../components/confirm-dialog/confirm-dialog.component';
+import { UpdateScheduleComponent } from './../../components/update-schedule/update-schedule.component';
+
 
 @NgModule({
   declarations: [
@@ -28,7 +33,12 @@ import { RegisterDialogComponent } from './../../components/register-dialog/regi
     ProfileComponent,
     UnloginComponent,
     LoginDialogComponent,
-    RegisterDialogComponent
+    RegisterDialogComponent,
+    ScheduleComponent,
+    NotifDialogComponent,
+    ConfirmDialogComponent,
+    UpdateScheduleComponent,
+
   ],
   imports: [
     CommonModule,
@@ -46,7 +56,12 @@ import { RegisterDialogComponent } from './../../components/register-dialog/regi
     ProfileComponent,
     UnloginComponent,
     LoginDialogComponent,
-    RegisterDialogComponent
+    RegisterDialogComponent,
+    ScheduleComponent,
+    NotifDialogComponent,
+    ConfirmDialogComponent,
+    UpdateScheduleComponent,
+
   ]
 })
 export class SharedModule { }
